Moved linalg functions from simple.c -> linalg.c and made corresponding header file changes
